Create Text to speech converter using python
Sometimes we prefer listening to the content instead of reading. We can do
multitasking while listening to the critical file data. Python provides many APIs to
convert text to speech. The Google Text to Speech API is popular and commonly known
as the gTTS API.
It is very easy to use the tool and provides many built-in functions which used to save
the text file as an mp3 file.
For using text-to-speech technologies offline well, pyttsx3 library (a text to speech
conversion library in Python) is used.
In this project, we have made a text to speech converter using this library.
Working Mode
1. User enters the text which is to be converted as speech.
2. Program provides an option for the user to check different voice rate:
 If user types yes, then different voice rates are demonstrated
 user selects the voice rate and hear the speech
 User is given another option to check different volumes
 If yes, then different volumes are demonstrated
 user selects the volume and hear the speech
 User can hear it in female voice if the user wishes to.
 After that audio file is saved.
 If user types no:
 User is asked to enter voice rate and volume
 User can hear it in female voice if the user wishes to.
 Audio file is saved.